Update for Django 1.5(beta1)
There were a couple of changes in the internals of Django which have
broken parts of django-refinery.

In django/django@d5a277ba4d608284991428fc3f7d6e0dbe6f8300 `QUERY_TERMS`
was changed to be a set. This is backwards compatible if you use the
fact iterating over a dictionary is the same as iterating over
`dict.keys()`.

In django/django@c4aa26a983c91b1ec015fe0552077847116dfab7 `LOOKUP_SEP`
was moved to a different module.

All the tests should now pass on Django1.4 and Django1.5b1.